requirements.txt
streamlit==1.32.0
streamlit-webrtc==0.47.0
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0
numpy==1.26.4
sounddevice==0.4.6
scipy==1.12.0
pydub==0.25.1
livekit-client==1.0.0
aiortc==1.5.0
flask==3.0.2 